+++ src/main/scala/ConstantString.scala
@@ -1,0 +1,5 @@
+package com.originate.scalypher
+
+abstract class ConstantString(stringified: String) {
+  def toQuery: String = stringified
+}
