+++ src/main/scala/PropertyReference.scala
@@ -1,0 +1,3 @@
+package com.originate.scalypher
+
+case class PropertyReference(name: String)
