+++ src/main/scala/ToQuery.scala
@@ -1,0 +1,11 @@
+package com.originate.scalypher
+
+import types._
+
+trait ToQuery {
+  def toQuery: String
+}
+
+trait ToQueryWithIdentifiers {
+  def toQuery(referenceableMap: ReferenceableMap): String
+}
